SSI-616: Resolve Spring container issues
The existing vulnerabilties and best practice problems do not apply to us,
either they are false positives, or minor issues we cant fix.
Updating to latest openj9 image.
Also enabling the vulnerability scan for the wallet. We might decide to drop this scan,
as the base image is not ours and we cant directly fix issues in that.