Update subrproject docker tasks
Task configurations come from the root build.gradle.
For Spring subproject we only need to configure the
ext properties with the correct image names.
The benefit is we do not need to separately configure
all the docker tasks in each subproject.

This restructure does not change the image building behaviour.
./gradlew buildAllImages still works the same.