Update to Acapy 0.8.0
Latest stable version.

See notes:
https://github.com/hyperledger/aries-cloudagent-python/releases/tag/0.8.0

There are several breaking changes.
The one effecting is the changing default webhook
message format.
Previously the default mode was the debug / verbose
mode. VCMS relies on this, expecting the presentation
exchange to contain the presentaiton itself.
This is a potential issue if the claims contain big
fields like images, increased the webhook message sizes.
In practice we did not see issues with that, for now
we can keep using the old behaviour.

Added an extra check in the startup script.
An explicit setting is expected for this new flag,
in practice for VCMS this has to be true right now.
We could also hardcode --debug-webhooks into the startup script.
Here Im opting for a more verbose config that is
also easier to configure and change later on. To be reviewed.