Support data uri input for claim value
The frontend might return a data uri as claim value.
The main use case is storing images, where we need to
set the correct mime type. The data uri is used
as a source to identify the mime type.

When a data uri is found the value is persisted as-is,
which normally means the base64 encoded value.
This needs to be compatible with the wallet apps.

To support backwards compatibility if data uri is not detected,
we just store the claim value as clear text.